Get into gameplay state (#22)
* Add some symbol maps to Rev0

* Complete the symbol map

* Add asset importers and get further into boot

* Actually call mainLoop

* Include model.h

* Add missing assets in symbol map (unused)

* Fix a bug in importer

* Fix a vtx desync

* The pointer truncationing begins

* Declare callers

* More fixing

* Moar

* Update code_0.c

* Stub this

* itemprint_init, welcome back

* Sprites misaligned

* More truncation

* More sprite alignment

* More truncation

* Maybe the last one?

* Update spawnqueue.c

* Enable comusic init and fix a few pointers

* Freelist

* The great truncationing fix

\

* Stub out things that try to use the defunct custom heap

* Init music system allocation and state, but don't let anything reach the synth driver

* Update os_convert.h

* Couple more truncations

* Update modelRender.c

* Update modelRender.c

* Truncations in model drawing

* More function prototype mess

* More truncation, infinite loop fix

* Fix ActorProp union, ALDMAproc, some audio ptr truncation

* Update model.h

* Fix ActorProp union, ALDMAproc, some audio ptr truncation

* Uncomment include in ultra64.h.
Comment includes necessary to run with aforementioned change.

* Last of the prototypes maybe

* Nope, not the last protos

* Fix infinite loop

* Try fixing a heap corruption with model loading

* Correct misclassified assets and log asset loads

* Holy crap it's at drawing

* Render to skybox

* Fix more truncations and some misc things

* Gather forward function declarations into a single prototypes.h file

* Forward declare AnimMtxList

* Enable overlays with static linking hook function

* Bump scratch size for sprite display data

* Read in TextureDataOffset

DL commands contain hardcoded segment 2 offsets from the original ROM layout. Previously, the port packed texture blobs contiguously, which shifted offsets whenever a texture occupied more ROM space than the extracted pixel data (e.g., RGBA16 textures loaded via G_LOADBLOCK span more bytes than width*height*bpp). Now each blob is placed at its original ROM position, so DL offsets match exactly.

* Fix sprite texture rendering via gDPSetTextureImage override

Raw BK sprite textures pass heap pointers through gDPSetTextureImage,
which crashes OTR signature checks. Add port_gDPSetTextureImage to
safely pass raw pointers and enable TMEM derivation for sprite tiles.

* Make render fill the window

* Move logging to debug

* Logging for cubes and cameras

* Adjust setup importer

* Boot directly to file select for diag

* Allow RelWithDebInfo

* Remove diag logging and ubstub glspline

GLspline sets up cutscene camera system, it's needed now

* More truncations

* Fill the display window

* The annoying opt-in we have to do until LUS #1004 is merged

* Fix animations a bit

* Update ultra64.h

* Boot into file select map for now

* Purge dead 2.0L header files

* Asset lookups now come from aBKAssetTable within the o2r

* Standardize c logging by wrapping spdlog in BK_LOG_* levels

* Fix more truncated pointers

* Adapt some filenames to reflect their purpose

* Dix more stuff including logic fps

* Cut log spam

* Fix sprite billboarding (needs Torch update)

* Functional controls

* Fix several truncated pointers and prototypes

* Show the current map id in the window title

* Import and apply correct sprite animation data (needs Torch update)

* Dix collision pointer truncations and segaddr tagging in models

* ifdef a few init things not used for safety

* Identified an animation

* Clean up and rename some files based on content

* Game logic needs to run independent of render logic

* Fix a glspline bug where an int declared as bool truncated the value

* Fix more 32->64 bugs

* Fix some audio truncations

* Give friendly names to all code_* files and create a directory  reference in docs/, eliminate cmake excludes

Most if not all code_* files have been given friendly, meaningful names which will help in trace logs. A table of contents has been created in the docs/ folder. CMakeLists no longer needs excludsion logic, all excludes were verified no longer needed for reference.

* Change references in BUILDING.md

* Add documentation and correct some comments, removbe deadcode from factory

* Update LUS submodule to JeodC/libultraship TMEM branch

* Fix ingame dialog conversations and hud sprites

* Correct more unknown assets by comparing to enum.h

* Unsure if this is correct, we'll find out. Detect between models and sprites.

* MASH B TO SKIP DIALOG LOL BYE BOTTLES

* Glasses toad lol sorry bottles Kazooie's idea

* Clang-format

* Name some dialog

* Revert "Name some dialog"

This reverts commit 5903b00f692ef230d64ed4cbfb7b7dae504ca3e2.

* Split text up into dialog, quizq and gruntyq

* Clang-format lus

* More 32-64bit fixes

---------

+++ include/2.0L/PR/region.h
@@ -73,8 +73,7 @@ extern "C" {
  */
 
 /* Perform alignment on input 's' */
-#define	ALIGN(s, align)	(((u32)(s) + ((align)-1)) & ~((align)-1))
-
+#define	ALIGN(s, align)	(((uintptr_t)(s) + ((align)-1)) & ~((uintptr_t)((align)-1)))
 
 /***************************************
  *
