+++ include/core2/timedfunc.h
@@ -8,11 +8,12 @@
 
 
 void timedFunc_set_0(f32 time, GenFunction_0 funcPtr);
-void timedFunc_set_1(f32 time, GenFunction_1 funcPtr, s32 arg0);
-void timedFunc_set_2(f32 time, GenFunction_2 funcPtr, s32 arg0, s32 arg1);
-void timedFunc_set_3(f32 time, GenFunction_3 funcPtr, s32 arg0, s32 arg1, s32 arg2);
-void timedFunc_set_4(f32 time, GenFunction_4 funcPtr, s32 arg0, s32 arg1, s32 arg2, s32 arg3);
-void timedFunc_set_5(f32 time, GenFunction_5 funcPtr, s32 arg0, s32 arg1, s32 arg2, s32 arg3, s32 arg4);
+// [port] s32 -> uintptr_t for 64-bit pointer safety
+void timedFunc_set_1(f32 time, GenFunction_1 funcPtr, uintptr_t arg0);
+void timedFunc_set_2(f32 time, GenFunction_2 funcPtr, uintptr_t arg0, uintptr_t arg1);
+void timedFunc_set_3(f32 time, GenFunction_3 funcPtr, uintptr_t arg0, uintptr_t arg1, uintptr_t arg2);
+void timedFunc_set_4(f32 time, GenFunction_4 funcPtr, uintptr_t arg0, uintptr_t arg1, uintptr_t arg2, uintptr_t arg3);
+void timedFunc_set_5(f32 time, GenFunction_5 funcPtr, uintptr_t arg0, uintptr_t arg1, uintptr_t arg2, uintptr_t arg3, uintptr_t arg4);
 void timedFunc_set_6(f32 time, GenFunction_6 funcPtr, void* argPtr );
 void timedJiggySpawn(f32 time, s32 jiggyId, f32 *position);
