+++ include/generic.h
@@ -2,11 +2,12 @@
 #define _GENERIC_METHODS_H_
 
 typedef void (* GenFunction_0)(void);
-typedef void (* GenFunction_1)(s32);
-typedef void (* GenFunction_2)(s32, s32);
-typedef void (* GenFunction_3)(s32, s32, s32);
-typedef void (* GenFunction_4)(s32, s32, s32, s32);
-typedef void (* GenFunction_5)(s32, s32, s32, s32, s32);
+// [port] s32 -> uintptr_t for 64-bit pointer safety
+typedef void (* GenFunction_1)(uintptr_t);
+typedef void (* GenFunction_2)(uintptr_t, uintptr_t);
+typedef void (* GenFunction_3)(uintptr_t, uintptr_t, uintptr_t);
+typedef void (* GenFunction_4)(uintptr_t, uintptr_t, uintptr_t, uintptr_t);
+typedef void (* GenFunction_5)(uintptr_t, uintptr_t, uintptr_t, uintptr_t, uintptr_t);
 typedef void (* GenFunction_6)(void *);
 
 #endif
